Move ThreadBackup from MainVM to ThreadVM
